Fix build for Qt 6.1
Task-number: QTBUG-89299
Task-number: QTBUG-90926
(cherry picked from commit d2c6249bb9292deec97b5a026be5efdb7a83bba7)

+++ tests/manual/barstest/chart.cpp
@@ -684,7 +684,9 @@ void GraphModifier::changeSelectionMode()
 {
     static int selectionMode = m_graph->selectionMode();
 
-    if (++selectionMode > (QAbstract3DGraph::SelectionItemAndColumn | QAbstract3DGraph::SelectionSlice | QAbstract3DGraph::SelectionMultiSeries))
+    if (++selectionMode > (int)(QAbstract3DGraph::SelectionItemAndColumn |
+                                QAbstract3DGraph::SelectionSlice |
+                                QAbstract3DGraph::SelectionMultiSeries))
         selectionMode = QAbstract3DGraph::SelectionNone;
 
     m_graph->setSelectionMode((QAbstract3DGraph::SelectionFlag)selectionMode);
